linuxrc: always copy conf files
Previous behavior was that when populating sysroot with the files inside
our conf directory, we create symlinks unless we were loading the
rootfs.img into RAM.  That was leading to config problems with some
applications that expect a file to be a normal file (e.g., /etc/sudoers)
and was also causing all kinds of regressions (e.g., what if the file
exists already in rootfs?  what if the config file you want to cp into
rootfs needs to actually be a symlink?).  So, regardless, we now copy
the files.

This should also alleviate the chicken-and-the-egg problem regarding the
injection of root-true into /etc/fstab when users have a modified version
in conf (see 7d8f6ea33cc8eff5889aba9e3cc06ff4e890f3b6).