linuxrc: use cp -dpR when copying conf files
Previously, we were doing this via find instead of using the recursive
behavior of cp.  I was trying to avoid having to sort out the gotchas I
knew would crop up if we used cp -R, but it tuns out there are quite a bit
more of them doing it via find... oops.

We still need a single find statement for dealing with the only real gotcha
that cropped up during testing.

NOTE: This also adds mv to the initramfs, which we somehow weren't using
      before this patch.