linuxrc: run fsck on root dev
This is needed because the root dev becomes impossible to unmount (and
thus, impossible to safely fsck) after we pass control over to the host os.
Without this, the only way to fsck the real root device is to load the
entire squashfs.img into RAM via the to_ram flag, but this might not even
be possible if the system has too little RAM or too large of a
squashfs.img.

NOTE: This also adds fsck, e2fsck, fsck.ext2, fsck.ext3, and fsck.ext4 to
      the initramfs.

NOTE: Unless we add other fsck.TYPE executables later, we ONLY support
      EXT2/3/4 filesystems.