made mdadm and lvm optional
This commit renames linuxrc to linuxrc.in and substitutes some configure
variables in to make both mdadm and lvm optional.  If a host system doesn't
have the lvm or mdadm binaries, it obviously can't be using LVM or RAID, so
we might as well not fail to build.  We now just disable the relevant
sections of the linuxrc boot script.