copy all progs into initramfs
Previsouly, we were trying to keep a subset of the progs installed on the
initramfs... but that was leading to duplicates and missing entries.  I've
gone and combined the two sections in NEEDED_PROGS and sorted the list
there and in ruckusrd.in.

This commit also adds rmmod, modinfo, and ps to the initramfs, because I
wished I had them the last time I was debugging.  ;-)