linuxrc: use symlink for /etc/mtab
This is being done to ensure that our mtab is correct during the initramfs
tasks.  If we use a symlink to /proc/mounts, we don't have to worry about
things getting out of wack.  I was running into remount failures on some
boxes when the original mount wasn't in /etc/mtab.

This commit also removes the -n from all mounts after we create /etc/mtab.