fighting with CentOS7
I can't get this to work with CentOS7's kernel or a custom kernel w/ mostly
everything built in...  I can't seem to get udevd to trigger
auto-modprobing of ahci (or anything, for that matter).

Furthermore, I know lvm support is borked because a bunch of errors come
out about not being able to talk to lvmetad... whatever that is.

- look for systemd-udevd and/or udevd and user whichever is found first
- cp all of both /etc/udev and /lib/udev
- copying modules.*
- copy ALL modules (not just boot-related ones) temporarily
- copy both .ko.xz and .ko files
- switched to xz compression of initramfs