brain dump of work that needs to be done by build/install system
Most of this was ripped right out of the ruckusrd script, some was jotted
down as I went along reviewing that and our linuxrc script.

Not sure if I want to do this via Make or via a script that gets invoked
via make... maybe some of both.  We probably do have some 'all' and some
'install' target work in here intermixed.