ruckusrd: use cpio instead of sqsh_layer for pregnerated initramfs img
Using an overlay mount of lower.sqsh worked... but now we've inadvertantly
made ruckusrd require root priviledges even when writing to a file the user
has write perms for.  That was never the intent.

This commit switches over to a cpio.xz file instead, which will probably
make invocations of ruckusrd take a smidge longer, but will work w/out
special priviledges.