ruckusrd: don't use chroot to prep udev hwdb
The previous commit removes the need for extra permissions regarding
loopback/overlay mounting... but we still need to be root to chroot into
the initramfs and prep the udev hwdb.  Oops.

This commit moves that work into linuxrc prior to starting udevd, which
should work just as well but will induce a teeny bit more work into the
bootup processing.