added doit script, v1
This utility script does some work in a spawned subprocess and emails the
user when complete w/ all the output of the job.

I've been using this for remotely kicking off kernel builds, so it
definitely works, but I'd call the script "not quite done" because I'm sure
it's full of lying comments from the initial implementation.  And it's not
plumbed into the build system yet.