linuxrc: look in both rootdev and sqsh_layerdev for layers
The most common use case for sqsh_layerdev is to share it read-only
amongst a group of virtual machines.  Allowing rootdev to be used IN
ADDITION to sqsh_layerdev, to augment it, allows those virtual
machines to have a read-write area to creat new layers in, w/out
forcing them to copy into sqsh_layerdev later and face potential
naming conflicts.

In other words, sqsh_layerdev is for SHARED layers and rootdev is for
PRIVATE layers.