ruckusrd: allow using symlink name for kernel version
If you're using kernel_builder's NOVERSION=stable or manually making
convenience symlinks to kernel versions, you can now use the symlink name
as the kernel version argument to ruckusrd.  This is specifically needed
so we can ensure that the specified version has a correlating config file
in /boot for our run-time checks.