ruckusrd: autodetect compression to use on initramfs
This commit adds a -c flag to ruckusrd to specify either "best" or "fast"
compression for the initramfs.  The supported compressors are zstd, lz4,
xz, and gzip.  At run-time, we check for the existence of the compression
tools on the host system and also for the correlating CONFIG_RD_WHATEVER in
/boot/config-VERSION.  The order we check for them is determined by the
"best" or "fast" compressor mode.  If we can't find one, we fall back to
gzip.  Default is "fast".