linuxrc: added fwdev commandline arg
This new argument is for providing a device to check for a fw.sqsh
file.  This image, if it exists, is mounted immediately upon initramfs
extraction so that any modules autoloaded by udev will find their
required firmware.

The resulting /lib/firmware directory is injected into the mounted
rootfs via modinject, if enabled.

Note that the fwdev argument must be an explicit device name (e.g.,
/dev/sda1), as this special bit happens prior to starting udev, which
is how we do our LABEL/UUID based device lookups.  (at least, for now)