added releases to ChangeLog
I have historically done this in the tag objects for each release.  I
guess it would be nice to have that info present in a source tarball,
so I've extracted it and put it all in here.  Plus, I seem to be the
only person on Earth putting release notes in tag objects...