subprojects/Makefile: fix less compilation
Not sure what's going on here, it seems like sometimes it compiles fine and
sometimes Make reports an error in the Makefile.  Seems like the error is
in the line that does a $(shell ${MAKE} echo_distfiles) call, which doesn't
seem needed...  I've just removed that line via a sed call.