linuxrc: redo the fwdev specification and mounting, v1
Well... this did work for systems where fwdev is a disk that's already
present, but using a USB disk for fwdev doesn't work because even
though we load all the modules, we don't wait for hotplug events to
finish.

This commit makes a bunch of changes, starting with changing the
fwdev= parameter to just take a single device like all the other
device arguments (why on earth was this ever a comma-delimited list of
potential devices?).

The strange loop that was iterating over potential devices has been
replaced with a wait-loop that attempts to locate the specified fwdev
at 1Hz for up to 10 seconds, allowing USB devices to get hotplugged.

We now also get some indication of success/failure, albeit pretty
quietly.  Failure is still treated as a warning, booting does not stop
(unless missing firmware actually causes a crash).