busybox: disable tc applet until upstream removes CBQ support
For building on noble (or anything that's using kernel headers newer
than v6.8), the tc module fails to build because CBQ support was
removed from the kernel.  It looks like this hasn't been fixed
upstream yet, even in the newer 1.37 release, so we'll disable tc for
now in our .config file.