debian: updated get_deps script
Had to fix this script for Ubuntu 24.04.  The new merged-usr setup
there was causing dpkg to print some extra words that needed to get
filtered out, because they're not package names.

We don't actually use the output of this script in ruckusrd, but I
want to make sure we have the latest version here so I can use
ruckusrd as the authoritative source of my dpkg autotools target.