linuxrc: added `initramsys-installer` boot param, v7
Ooh, really starting to make progress.  Big change here was plumbing
things up so I can actually put all the files on a single piece of
boot media, instead of supplying a vm with separate disks for root and
imgdev (which i only ever did becuase it was convenient for testing).

- Can now specify initramsys-installer w/out any arguments to just
  take sane defaults.
- Split swap ammount by number of disks.
- Added skeletal efibootmgr call.
- Lots of comment cleanup and additional FIXMEs