linuxrc: really wait for /dev/root
I'm not sure how long this has been broken for...  but it definately
was.  We were only entering the /dev/root wait if /dev/root was a
symlink, and then we were waiting for what it points at to
exist... but that symlink doesn't actually get created by udev until
the device DOES exist... so there's not a symlink at all and we don't
wait...  behaviorly completely wrong, but I know this did work years
and years ago when I first wrote this.  Odd.