linuxrc: fix /dev/root mount failure recovery handling
This is another "used to work, wonder how long it's been broken"
problem.  We store the return value of the mount command in a
variable, then check to see if non-zero and drop to a control shell
with a helpful message telling the user how to fix and continue
booting... except we trap all errors so that code no longer gets
executed.

Shrunk the handling code up into a simple || error handling block to
allow the helpful code to execute.