linuxrc: really wait for /dev/root, v2
We still have some corner cases here, most notably we were checking to
see if /dev/root already exists before going into the wait code...
but we *need* to go in there to figure out fstype.  So this takes out
that conditional entirely and treats rootdev and sqsh_layerdev the
same.