debian: fixed get_deps flagging our own package as a dependency
This commit fixes the corner-case where a built .deb file ends up
depending on itself because there was a previous version of a library
that's provided by this package intalled on the system.

This doesn't necesarily effect ruckusrd, but this is the repo I keep
my master dpkg build rules in.