linuxrc: consistently mount/umount all virtual filesystems including efivars
This commit ensures that we've got ALL the virtual filesystems mounted
any time we mount any of them.  This is slightly non-trivial because
it's entirely possible we're booted under BIOS which makes the efivars
mount fail.

We can now expect to have a fully funcional chroot system in maint,
initramsys (after running mount_sysroot), initramsys-installer, and
firstboot_wizard.