system_installer: split can_efi into create_esp and modify_efivars
Trying to use can_efi to mean two different things depending on other
variables/situations we getting messy and error-prone.  This commit
has us set can_efi and create_esp depending on the value passed in via
EFI= and cleans things up a bit.

In so doing, we've changed the behavior of swap creation slightly.
Previously, if autodecting can_efi failed, swap would also be
dislabled.  The reasoning was that I liked that behavior, despite it
being unrelated in the real world, because I was testing on domUs that
didn't have EFI support and cooincidentally I didn't want swap
created.  But that was just lazy, and of course, now obviously wrong
(e.g., if installing when boooted from BIOS, we would disalbe swap
too? that's wrong).

We were also not creating swap when EFI=default (to use default EFI
boot paths and not modify EFIVARS), because that implies you're
creating bootable removable media... which was also an assumption.

To disable swap, you must now explicitly specify SWAP=0 on the command
line.

Also renamed want_swap -> create_swap for consitency.