linuxrc: tell udhcpc to run in background so initramsys-net is safe
Previously, in initramsys-net, udhcpc would fail emergency_shell if no
link or would wait forever if link but no dhcp response.  Not good.
We now use -b instead of -f to tell it to go into the background and
wait for a response.  Much better.  Now initramsys-net is safe to
always use.